Fix potential bugs when Math.abs(Integer.MIN_VALUE)<0
It is possible for a call to hashCode to return Integer.MIN_VALUE.
Take the absolute value of such a hashcode and you'll still have a negative number.
To fix the problem, cast it to long.

+++ src/main/java/groovy/lang/GroovyClassLoader.java
@@ -270,7 +270,7 @@ public class GroovyClassLoader extends URLClassLoader {
      */
     public Class parseClass(String text) throws CompilationFailedException {
         return parseClass(text, "script" + System.currentTimeMillis() +
-                Math.abs(text.hashCode()) + ".groovy");
+                Math.abs((long) text.hashCode()) + ".groovy");
     }
 
     public synchronized String generateScriptName() {
